Reorganize dev source files.
Also move components under src/apps/details/.
