+++ react-components/config/paths.js
@@ -34,7 +34,8 @@ module.exports = {
   appBuild: resolveApp('../de-webapp/src/main/webapp/react_components'),
   appPublic: resolveApp('public'),
   appHtml: resolveApp('public/index.ejs'),
-  devHtml: resolveApp('public/index.html'),
+  devHtml: resolveApp('public/dev/index.html'),
+  devDir: resolveApp('public/dev'),
   appIndexJs: resolveApp('src/index.js'),
   appLibJs: resolveApp('src/lib.js'),
   appPackageJson: resolveApp('package.json'),
