+++ react-components/public/dev/apps/details/category_tree.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Category Tree</title>
+  </head>
+  <body>
+    <h1>Category Tree</h1>
+    <div id="category-tree"></div>
+  </body>
+</html>
