+++ react-components/public/dev/apps/details/copy_text_area.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Copy Text Area</title>
+  </head>
+  <body>
+    <h1>Copy Text Area</h1>
+    <div id="copy-text-area"></div>
+  </body>
+</html>
