+++ react-components/public/dev/apps/details/tool_info.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Tool Info</title>
+  </head>
+  <body>
+    <h1>Tool Info</h1>
+    <div id="tool-info"></div>
+  </body>
+</html>
