+++ react-components/public/dev/index.html
@@ -16,14 +16,11 @@
     <title>React App</title>
   </head>
   <body>
-    <h1>Tool Info</h1>
-    <div id="tool-info"></div>
-
-    <h1>Copy Text Area</h1>
-    <div id="copy-text-area"></div>
-
-    <h1>Category Tree</h1>
-    <div id="category-tree"></div>
+    <ul>
+      <li><a href="apps/details/category_tree.html">apps/details/category_tree.html</a></li>
+      <li><a href="apps/details/copy_text_area.html">apps/details/copy_text_area.html</a></li>
+      <li><a href="apps/details/tool_info.html">apps/details/tool_info.html</a></li>
+    </ul>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
