+++ react-components/src/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './CopyTextArea';
+import App from './apps/details/CopyTextArea';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
