+++ react-components/src/apps/details/CategoryTree.js
