+++ react-components/src/apps/details/ToolDetails.js
